Create system for spawning projectile effects and despawning them after one animation cycle.
