+++ thetawave-projectiles/src/attributes/mod.rs
@@ -1,7 +1,11 @@
 mod data;
 mod plugin;
 
-pub(crate) use data::{ProjectileAttributesResource, ProjectileRangeComponent};
+pub(crate) use data::{
+    DespawnAfterAnimationComponent, ProjectileAttributesResource, ProjectileRangeComponent,
+};
 pub(crate) use plugin::ThetawaveAttributesPlugin;
 
-pub use data::{ProjectileSpawner, ProjectileType, SpawnProjectileEvent};
+pub use data::{
+    ProjectileSpawner, ProjectileType, SpawnProjectileEffectEvent, SpawnProjectileEvent,
+};
